removed deprecated mvn and fixed int in linspace

+++ dual_process.py
@@ -29,7 +29,7 @@ EPS         = 1e-10 # a very small value (used for numerical stability)
 NR_THREADS  = 1;    # this is for multithreaded fft
 DELTA_T     = 0.025;  # size of discrete time increment (sec.)
 MAX_T       = 8.0; #ceil(percentile(all_RT,99.5))
-NR_TSTEPS   = MAX_T/DELTA_T;
+NR_TSTEPS   = int(MAX_T/DELTA_T);
 NR_SSTEPS   = 8192;
 NR_SAMPLES  = 10000; # number of trials to use for MC likelihood computation
